half a dozen new indices for ranking
could probably get away with either just ranking records (upper bound for DIDs) or DIDs (lower bound for records) but ordering by the other would probably be messy / approximate. these indices should be pretty cheap to store, so doing it is must a matter of overcoming my laziness

to that effect i'm finally adding a macro for the static_str thing, and made the keys generic-ish.